refactor: streamline TaskDetailsDialogActionButtons and remove custom hook
- Refactored TaskDetailsDialogActionButtons to directly manage task actions, eliminating the custom hook for improved clarity and maintainability.
- Enhanced prop handling by defining the interface directly within the component, simplifying the structure.
- Updated task action handlers to provide better error handling and user feedback during task updates and deletions.
- Removed legacy code related to Privy authentication, ensuring a more straightforward implementation of task actions.
