+++ components/VercelChat/dialogs/tasks/TaskDetailsDialogActionButtons.tsx
@@ -2,21 +2,80 @@
 
 import { Button } from "@/components/ui/button";
 import { Pause, Trash2 } from "lucide-react";
-import { usePrivy } from "@privy-io/react-auth";
-import {
-  useTaskDetailsDialogMutationHandlers,
-  type TaskDetailsDialogActionButtonsProps,
-} from "./useTaskDetailsDialogMutationHandlers";
+import { useUpdateScheduledAction } from "@/hooks/useUpdateScheduledAction";
+import { useDeleteScheduledAction } from "@/hooks/useDeleteScheduledAction";
 
-export type { TaskDetailsDialogActionButtonsProps } from "./useTaskDetailsDialogMutationHandlers";
+export interface TaskDetailsDialogActionButtonsProps {
+  taskId: string;
+  editTitle: string;
+  editPrompt: string;
+  editCron: string;
+  editModel: string;
+  onSaveSuccess: () => void;
+  onDeleteSuccess: () => void;
+  isEnabled: boolean;
+  canEdit: boolean;
+}
 
 const TaskDetailsDialogActionButtons: React.FC<
   TaskDetailsDialogActionButtonsProps
-> = (props) => {
-  const { authenticated } = usePrivy();
-  const { handleDelete, handlePause, handleSave, isLoading } =
-    useTaskDetailsDialogMutationHandlers(props);
-  const { canEdit, isEnabled } = props;
+> = ({
+  taskId,
+  editTitle,
+  editPrompt,
+  editCron,
+  editModel,
+  onSaveSuccess,
+  onDeleteSuccess,
+  isEnabled,
+  canEdit,
+}) => {
+  const { updateAction, isLoading: isUpdating } = useUpdateScheduledAction();
+  const { deleteAction, isLoading: isDeleting } = useDeleteScheduledAction();
+  const isLoading = isUpdating || isDeleting;
+
+  const handlePause = async () => {
+    if (!canEdit) return;
+    try {
+      await updateAction({
+        updates: { id: taskId, enabled: !isEnabled },
+        successMessage: isEnabled ? "Task paused" : "Task activated",
+      });
+    } catch (error) {
+      console.error("Failed to toggle task status:", error);
+    }
+  };
+
+  const handleDelete = async () => {
+    if (!canEdit) return;
+    try {
+      await deleteAction({
+        actionId: taskId,
+        onSuccess: () => onDeleteSuccess(),
+      });
+    } catch (error) {
+      console.error("Failed to delete task:", error);
+    }
+  };
+
+  const handleSave = async () => {
+    if (!canEdit) return;
+    try {
+      await updateAction({
+        updates: {
+          id: taskId,
+          title: editTitle,
+          prompt: editPrompt,
+          schedule: editCron.trim(),
+          model: editModel,
+        },
+        onSuccess: () => onSaveSuccess(),
+        successMessage: "Task updated successfully",
+      });
+    } catch (error) {
+      console.error("Failed to save task:", error);
+    }
+  };
 
   return (
     <div className="flex gap-2 mt-4 pt-4 border-t border-border justify-between shrink-0">
@@ -30,18 +89,16 @@ const TaskDetailsDialogActionButtons: React.FC<
           <Pause className="h-4 w-4 mr-2" />
           {isEnabled ? "Pause" : "Resume"}
         </Button>
-        {authenticated && (
-          <Button
-            variant="outline"
-            onClick={handleDelete}
-            className="border-red-200 text-red-600 hover:bg-red-50"
-            disabled={isLoading || !canEdit}
-            size="sm"
-          >
-            <Trash2 className="h-4 w-4 mr-2" />
-            Delete
-          </Button>
-        )}
+        <Button
+          variant="outline"
+          onClick={handleDelete}
+          className="border-red-200 text-red-600 hover:bg-red-50"
+          disabled={isLoading || !canEdit}
+          size="sm"
+        >
+          <Trash2 className="h-4 w-4 mr-2" />
+          Delete
+        </Button>
       </div>
       <Button onClick={handleSave} disabled={isLoading} size="sm">
         {isLoading ? "Saving..." : "Save"}
